added static otp for dealer mapping
